Fix compiler warning: implicit conversion loses integer precision
grisbi/src/gsb_transactions_list.c:259:9: warning: implicit conversion loses integer precision: 'glong' (aka 'long') to 'gint' (aka 'int') [-Wshorten-64-to-32]
  259 |                 len = g_utf8_strlen (text, -1);
      |                     ~ ^~~~~~~~~~~~~~~~~~~~~~~~

+++ src/gsb_transactions_list.c
@@ -247,7 +247,7 @@ static void transactions_search_entry_icon_released (GtkEntry *entry,
 static void gsb_transactions_list_update_search_filter (void)
 {
 	gint account_number;
-	gint len = 0;
+	glong len = 0;
 	gint matches = 0;
 	const gchar *text = NULL;
